Rename which_exec -> which_function
