+++ tests/lib.rs
@@ -119,7 +119,7 @@ mod timestamps;
 mod undefined_variables;
 mod unexport;
 mod unstable;
-mod which_exec;
+mod which_function;
 #[cfg(windows)]
 mod windows;
 #[cfg(target_family = "windows")]
